Change the hook
The previous hook got called 'before' the access dates were set. So it was not possible to check the access date and decide if the email should be sent or not. This hook is called after the date is set. So it's possible to send the email conditionally by checking the access start date